+++ dormitory_management/src/main/java/com/example/dormitory_management/entity/Contract.java
@@ -16,7 +16,7 @@ public class Contract {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "contract_id")
-    private Integer contractId;
+    private Long contractId;
 
     @ManyToOne(fetch = FetchType.LAZY, optional = false)
     @JoinColumn(name = "student_id", nullable = false)
