+++ dormitory_management/src/main/java/com/example/dormitory_management/entity/RoomPrice.java
@@ -17,7 +17,7 @@ public class RoomPrice {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "price_id")
-    private Integer priceId;
+    private Long priceId;
 
     @ManyToOne(fetch = FetchType.LAZY, optional = false)
     @JoinColumn(name = "room_id", nullable = false)
